pyproject.toml
[project]
name = "repl-nix-publicrisknavigator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.58.1",
    "pdfkit>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
    "trafilatura>=2.0.0",
    "weasyprint>=63.1",
    "werkzeug>=3.1.3",
]
